rename: styles폴더명 수정 (#30)
